remove redundant fixture
